made code look a bit better

+++ 2025Robot/src/main/java/frc/robot/subsystems/Field.java
@@ -152,7 +152,7 @@ public class Field extends SubsystemBase {
                 poseSetpoint = new Pose2d(7.29, 6.18, Rotation2d.fromDegrees(0));
                 break;
             case Climb:
-                switch ((int) driverLocation.getSelected()) {
+                switch (driverLocation.getSelected()) {
                     case 1:
                         poseSetpoint = new Pose2d(7.33, 7.27, Rotation2d.fromDegrees(0));
                         break;
@@ -183,7 +183,7 @@ public class Field extends SubsystemBase {
      * Gets the nearest zone to the robot's current position.
      * @param robotPose -The robot's current position.
      * @return -Nearest zone to the robot from "zoneMap".
-     * @see -Link to nearest method: https://github.wpilib.org/allwpilib/docs/release/java/edu/wpi/first/math/geometry/Pose2d.html#nearest(java.util.List).
+     * @see -Link to "nearest" method: https://github.wpilib.org/allwpilib/docs/release/java/edu/wpi/first/math/geometry/Pose2d.html#nearest(java.util.List).
      */
     public Pose2d getNearestZone(Pose2d robotPose){
         return robotPose.nearest(zoneMap);
